Extract the RateLimit module (login throttling) into Api/RateLimit
Fifth module of the #777 folder migration. Moves the login-throttling
primitives into Jellyfin.Plugin.SSO_Auth.Api.RateLimit:

- SsoRateLimiter, PerClientBudgetLimiter, IntervalGate, ProviderScopedKey ->
  Api/RateLimit/ (namespace Api.RateLimit), import added at each call site.

RateLimit depends on Net and only Net (SsoRateLimiter keys its buckets by the
Net client-IP classifier). Registers the RateLimit -> Net edge on the module
DAG rule.

The DAG rule did its job: it caught the Net leaf IpAddressClassifier gaining an
Api.RateLimit import via two <see cref="SsoRateLimiter.NormalizeClientKey"/> doc
references. Those crefs are dropped to plain-text mentions so the Net leaf stays
decoupled (same treatment the Avatar move gave its AvatarUrlValidator cref).

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1576 both TFMs), the module
DAG conformance rule, and the ABI-floor build.

Refs #777
