+++ SSO-Auth.Tests/KeyedLockStoreTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.RateLimit;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
