+++ SSO-Auth.Tests/PerClientBudgetLimiterTests.cs
@@ -1,6 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.RateLimit;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
