+++ SSO-Auth.Tests/ProviderScopedKeyTests.cs
@@ -1,5 +1,6 @@
 using System;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.RateLimit;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
