+++ SSO-Auth.Tests/SsoRateLimitGateTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Globalization;
 using System.Net;
 using Jellyfin.Plugin.SSO_Auth.Api.Shared;
+using Jellyfin.Plugin.SSO_Auth.Api.RateLimit;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
