+++ SSO-Auth.Tests/SsoRateLimiterTests.cs
@@ -3,6 +3,7 @@ using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.RateLimit;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
