+++ SSO-Auth/Api/ProviderNameValidator.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Buffers;
 using Jellyfin.Plugin.SSO_Auth.Api.Net;
+using Jellyfin.Plugin.SSO_Auth.Api.RateLimit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Api;
